MIDI Meta channels can be up to channel 255, so do not reserve 4-bits but 8-bits for channels

+++ midi/midi.cpp
@@ -776,7 +776,7 @@ MIDI::new_channel(uint8_t channel_no, uint16_t bank_no, uint8_t program_no)
 }
 
 MIDIChannel&
-MIDI::channel(uint8_t channel_no)
+MIDI::channel(uint16_t channel_no)
 {
     auto it = channels.find(channel_no);
     if (it != channels.end()) {
